requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
google-api-python-client==2.141.0
google-auth==2.33.0
pytz==2024.1